feat(phase-7): decimal units, coupon/campaign engine, resumable stocktaking, real diagnostics, offline sync queue, local image storage
- units: Numeric(14,3) quantities everywhere; unit metadata (allow_decimal/decimals);
  fractional qty rejected for indivisible units; kg/gram/liter seeded
- marketing: Campaign/Coupon/CouponRedemption; percent+fixed, min purchase,
  max discount cap, usage limit, customer-specific, next-purchase auto-issue;
  consumed atomically inside the checkout txn (a failed sale never burns a coupon)
- stocktaking: warehouse, cursor/resume, counted_at/by, decimal counts, paged
  mobile item feed, bulk offline replay with per-row rejection reporting
- diagnostics: Connection Center performing REAL I/O per check (db write/read/
  rollback, storage probe, LAN address, provider queries, printer sinks);
  hardware without a device is SKIPPED, never PASS
- sync: durable SyncJob queue with backoff + idempotency; invoice SMS routed
  through it so connectivity never blocks a sale
- resolvers: images are downloaded, dimension-validated and stored under
  MEDIA_DIR; only the local path is referenced
- POS: search by name/barcode/SKU/model, coupon at checkout, phone-only customer
- settings: explicit IRT/IRR base currency (no silent conversion)
- 32 new tests (120 total, all passing)

+++ backend/app/routers/batches.py
@@ -20,7 +20,7 @@ router = APIRouter(prefix="/batches", tags=["batches"])
 class ReceiveIn(BaseModel):
     product_id: int | None = None
     barcode: str | None = None
-    quantity_received: int = Field(gt=0)
+    quantity_received: Decimal = Field(gt=0)  # decimal-aware (§25)
     buy_price: Decimal = Field(ge=0)
     consumer_price: Decimal | None = None
     sell_price: Decimal | None = None
@@ -33,7 +33,7 @@ class ReceiveIn(BaseModel):
 def _out(b: ProductBatch) -> dict:
     return {
         "id": b.id, "product_id": b.product_id, "batch_number": b.batch_number,
-        "quantity_received": b.quantity_received, "current_qty": b.current_qty,
+        "quantity_received": float(b.quantity_received), "current_qty": float(b.current_qty),
         "buy_price": float(b.buy_price), "consumer_price": float(b.consumer_price),
         "sell_price": float(b.sell_price),
         "production_date": str(b.production_date) if b.production_date else None,
